Fixed pending tests of completed features
Some tests were pending because they were failing although the
feature itself was working correctly. I've made some improvements
to these tests and made them work. The reason for failing was
mainly due to not waiting for ajax requests to complete and there
was a mistake in specifying the checkbox variable.